detached maplayerbutton into its own separate element for easier re-use
